dockerized/python-script/requirements.txt
numpy
matplotlib
scipy
kafka-python
faust
